Initial commit
I've been working on this project for about a month now and it's getting
ready for its very first release. I thought now that this is at least
doing something, I should finally do this commit. It is still missing a
few features that I want before showing this to anyone, but there is not
much missing.